// src/Verisoft.Frontend/src/libs/demo/documentation/ui/src/lib/models/constants.model.ts
export const FEATURE_TAB_NAME = 'feature';
export const API_TAB_NAME = 'api';

export enum DOCS_BLOCK_TYPE {
  SECTION = 'SECTION',
  API_SECTION = 'API_SECTION',
  API_ITEM = 'API_ITEM',
  CODE = 'CODE',
  TEXT = 'TEXT',
  TABLE = 'TABLE',
}

export enum DOCS_CODE_LANG {
  TS = 'ts',
  JS = 'javascript',
  CSHARP = 'csharp',
  HTML = 'html',
  JSON = 'json',
  UNKNOWN = 'text',
}